Move Concept specific controllers out of properties into concept
This is a first step to more clearly represent the **current** hirachical structure.

This will hopefully in the future make it simpler to actually create "shared" properties classes.

This PR does not move out History / Hierarchy Controllers. Since they currently are not exclusivly used
by the ConceptController, but also still part of kview.lidr.mvvm.view.properties.
