+++ kview/src/main/resources/dev/ikm/komet/kview/mvvm/view/concept/edit-descriptions.fxml
