+++ kview/src/main/resources/dev/ikm/komet/kview/mvvm/view/concept/edit-fully-qualified-name.fxml
